requirements.txt
absl-py==1.0.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==36.0.0
docutils==0.18
facenet-pytorch==2.5.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.8.2
jeepney==0.7.1
keyring==23.2.1
Markdown==3.3.6
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634106693478/work
oauthlib==3.1.1
olefile==0.46
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pkginfo==1.8.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
readme-renderer==30.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
SecretStorage==3.3.1
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.10.0
torch-tb-profiler==0.3.1
torchaudio==0.10.0
torchvision==0.11.1
tqdm==4.62.3
twine==3.6.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
zipp==3.6.0
